Reduce the number of times that test workers are created
Summary:
We were creating test workers every time we edited code
in the validation tests tab, despite the fact that there
are syntax errors much of the time. This changes it so we don't
attempt to go into that worker logic if there are syntax errors
in the test code.

Test Plan:
- Edit a challenge, go to validation tests tab
- Write tests for a few minutes, don't see it crash (and don't see lots of test-worker NET requests happen)

Reviewers: brian

Reviewed By: brian

Subscribers: #cs

Differential Revision: http://phabricator.khanacademy.org/D12398