Disable local completer for now
The local completer has some known issues like if you do things like:
this.x-3, it will think `this.x-3` is an identifier.

I fixed most of those inconsistencies with the local completer by wrapping it, but I'd rather ship it without any known issues like that.
Most of the value in the autosuggest is in the live documentaton.  I'd rather enable the local completer once I re-write it using the abstract syntax tree.

Test Plan:
- Test that autosuggest works.
- Test that no autosuggest happens for locally defined things.

Auditors: john