Fixed problem with talkie due to regex check in hotnumber
There was a regex change added here:
https://github.com/Khan/live-editor/commit/c75dced73d621f23272321fc39f7cbf28028c6b6#diff-ab911d7a1d7cf17b850e511c16b1a796L657

So that it only entered that condition when there was a number.
This changes it so that it goes back to the normal else { ... } condition without doing an additional test there.
The code for the autosuggest is inside that else and only used if the condition here:
https://github.com/Khan/live-editor/commit/c75dced73d621f23272321fc39f7cbf28028c6b6#diff-ab911d7a1d7cf17b850e511c16b1a796L660
is not hit.

Test Plan:
- Play ohttp://stable.ka.local/computing/cs/programming/variables/p/variable-expressions around -5:40 and make sure faceSize doesn't show up messed up.