Don't show any of the hotnumber scrubbers if within a string or comment
Summary:
I don't *think* this will cause problems with talkies because the change only comes into effect if you're inside a line that starts with / or * or within a string.
I think we'd avoid those things as to not show the pickers in talkies because it's a bug.

Fixes T3036.

Test Plan:
- Try to get autosuggestion, param info, color picker or number scrubber within comments and strings, should not work.
- Try to get the same outside of comments and strings, it should still work.

Reviewers: john, pamela

Reviewed By: pamela

Maniphest Tasks: T3036

Differential Revision: http://phabricator.khanacademy.org/D12502