Don't show draw canvas upon load
Summary:
I'm not sure what we were previously doing (pre-live-editor),
but we don't want to show the draw canvas upon load.
Instead, we should only show it when draw commands start (drawStarted)
and then hide it when draw commands end (drawEnded).
This fixes the issue with users unable to interact (T3046).

Test Plan:
- Watch Modifying arrays, confirm you can press to see balloons at 00:37
- Watch Intro to Drawing, confirm you can see the drawings show up at 5:00

Reviewers: john

Reviewed By: john

Subscribers: #cs

Differential Revision: http://phabricator.khanacademy.org/D12638