We don't need the horizontal scrollbar in the ace_scroller area. Fixes T3096.
Summary: I fiddled with the CSS on the main page in the Chrome console and changing this property did the trick. Seems rather painless to me!

Test Plan: Uhh - should be ok!? I tested editing code in the editor and it seemed to work fine for me.

Reviewers: pamela

Reviewed By: pamela

Maniphest Tasks: T3096

Differential Revision: http://phabricator.khanacademy.org/D12816