Add in HTML Screenshots using html2canvas. Add Webpage tests. Improve execution of tests.
Summary:
This adds in screenshots to the webpage type in Live Editor. It uses the html2canvas package, which works fairly well (there are some inconsistencies, but no show stoppers).

Additionally I added in a getScreenshot test for webpage and pjs output.

I also re-organized the tests a bunch and got it running smoothly for both types.

Test Plan: Run the tests, they all pass!

Reviewers: pamela, brian

Reviewed By: brian

Subscribers: jlfwong, #cs

Differential Revision: http://phabricator.khanacademy.org/D12955