Fix the height of the HTML output.
Summary: This fixes the height of the HTML output, it was defaulting to the default iframe height. It's weird as this only seems to happen on the main KA site, but this fixes that anyway (with no adverse side effects).

Test Plan: I confirmed that this change worked by changing this CSS property in the Chrome console on the live site.

Reviewers: brian, pamela

Reviewed By: pamela

Differential Revision: http://phabricator.khanacademy.org/D12978