Implementing HTML linting using Slowparse.
Summary:
This integrates the Mozilla Slowparse implementation to lint the user-generated HTML. Errors are fed into the normal "Oh Noes!" guy flow.

A couple points:

 - Like with JS, if there is a fatal error the HTML is not "executed" (injected into the page).
 - Error messages are derived from the errors that Slowparse/Thimble uses (heavily modified, had to re-word them in a number of places).
 - I had to write additional JavaScript-specific tests as Slowparse doesn't ship them by default (banning the script tag, banning event handlers, banning javascript: urls).
 - I integrated most of Slowparse's test suite into ours and included additional tests for the new JavaScript-specific stuff.

Sample:
https://www.dropbox.com/s/nefqammba38rwyz/Screenshot%202014-09-09%2012.07.41.png?dl=0

Test Plan:
Run the test suite, it all works!

Additionally attempt to generate an error in the editor itself. Note that the "Oh Noes"/Tipbar pops up with the error, the error is nicely formatted, and it includes useful variables. Make sure that the "Show me" link also jumps to the right position in the code.

Reviewers: pamela, brian

Reviewed By: brian

Subscribers: jlfwong, #cs

Differential Revision: http://phabricator.khanacademy.org/D13011