Make sure that mouse events are deleted
Summary:
Fixes T2723.

We had a test for checking if mouse events were deleted before,
but that test wasn't quite checking that, because it was calling
the relevant mouse function itself, and doing so meant the code
went down a different path where the mouse function would get replaced.
So we needed a test that would simulate what the user actually experiences,
where they delete the function and don't call it themselves, but instead
use their mouse to interact with the canvas.
I tried many things, but in the end, I think this is the best approach-
it actually simulates a click event on the canvas between tests.

For the actual fix, I changed a line that was only checking for 'draw'
to check for all the loop-inducing methods instead.

Test Plan:
- Run output_test.js

- Make a program like this:

```
var mouseMoved = function() {
    rect(mouseX, mouseY, 10, 10);
};
```
Move mouse and see the boxes appear.
Change program to this:
```
var xmouseMoved2 = function() {
    rect(mouseX, mouseY, 10, 10);
};
```
Move mouse, and still see the boxes appear.
Remove code, boxes sitll appear.

Reviewers: john

Reviewed By: john

Subscribers: #cs

Maniphest Tasks: T2723

Differential Revision: http://phabricator.khanacademy.org/D13336