Multi-recorder worker was loaded from external, not workers.
Summary: This was a mess up from the earlier commits, the mutli recorder worker was being loaded from external, which is not provided on the kasandbox server. Caught this on deploy. This should fix that!

Test Plan: Verify that attempting to do a new recording works. Hopefully this works on staging, as well!

Reviewers: brian, pamela

Reviewed By: pamela

Subscribers: #cs

Differential Revision: http://phabricator.khanacademy.org/D13984