Add some SQL output tests
Summary:
As you can see in the gulpfile, phantomJS doesn't have support for typed arrays,
So SQL tests must be loaded in a browser manually for now if you want to run them.

I'd also like to improve the tests to validate the types of errors being thrown,
but for now I just want to get some tests in.  Most of the test work was in the linting tests.

Test Plan: - Run tests by loading index.html in a browser

Reviewers: pamela, john

Subscribers: #cs

Differential Revision: http://phabricator.khanacademy.org/D14245