Fix matching SQL table results for more than 1 step with differing results
If step 1 of an SQL challenge had a result and you check for matching results, everything is ok.
But then on step 2, the template has 2 selects and 2 reuslts. So the step2 would pass, but the step 1
would fail because you had an extra result.

So the fix is just to allow more results and only check for the number of results in the template tables.

Test Plan:
- Test a challenge with 2 steps which both check results.

Auditors: pamela