Catch slowparse errors and spit out generic error
Summary:
We discovered that Slowparse.HTML can throw an error at times
when its parsing strategy doesn't work (like this particular case,
 where an unclosed attribute happens to have an http in it that
 looks like an attribute name and confuses it).
Ultimately we could improve the parser for the cases we found,
but in meantime, we should give the user some sort of error.
Note that we tried the same things on Thimble and saw it just doesn't
spit out any error at all when Slowparse.HTML fails.
This approach seems better!

Test Plan:
- Run output_test in webpage
- Start writing a link in https://www.khanacademy.org/devadmin/content/preview/x93850788/p/challenge-links-you-love,
   see the new error pop up after writing the first quote.

Reviewers: alexrodrigues

Reviewed By: alexrodrigues

Subscribers: #cs

Differential Revision: http://phabricator.khanacademy.org/D14502