Always process and send back tests - fixes T3361
Summary:
For some reason our HTML test logic differed from our PJS test logic in that
we didn't try calling .test() if there were errors, and we called back
with an empty set for test results.
However, that meant the challenge panels didnt know what tests existed.
I removed that if condition and the challenges seem to work as expected.
I'm not sure if there was a particular reason for the condition that I'm missing.

Test Plan:
- Go in an HTML challenge, make a syntax error, see the Steps remain above,
instead of it going into Completion mode

Reviewers: john, alexrodrigues, brian

Reviewed By: brian

Subscribers: brian, #cs

Differential Revision: http://phabricator.khanacademy.org/D14511