Hooked up remove()
Summary:
We were having problems because tooltips were not being correctly cleaned up when someone called remove() on live-editor. This lead to the selectors used by bootstrap tab targetting the wrong window in image-modal and that lead to display errors. The removes and associated re-initalizing is all because of the backbone router and so this issue only appears when the backbone router has changed pages at least once.
I threaded together the remove calls so that when you call remove() on live-editor remove() is called on all the tooltips.

Test Plan: Added test case in tooltips_test.js

Reviewers: john, pamela

Reviewed By: pamela

Subscribers: #cs

Differential Revision: http://phabricator.khanacademy.org/D14510