Add target="_blank" to all links
We had an issue where clicking on a link did actually
attempt to replace the frame, due to the link having same target
and the browser executing that before the onlick seemingly,
and that resulted in DOM security errors that persisted in playback.
This adds target=_blank to all links so that security error
does not happen.
(We could also handle security errors better if we run into
others in the future, to not break experience as badly)

Test Plan:
- Watch https://www.khanacademy.org/computing/computer-programming/html-css/html-tags-continued/p/html-links
 until the point where the link is inserted. Pause and click the link.
- Continue playing- you should *not* see an Oh Noes error.

Auditors: brian

Subscribers: alexrodrigues, #cs