Sound functionality and picker for PJS programs
Summary:
This adds a playSound() and getSound() functions
which allow students to play sounds in their programs,
plus a tooltip and picker for getSound.

The functions themselves are simple and just use the
HTMLAudioElement API. The support is pretty good for <audio> with MP3
across browsers, and this is a bonus feature, so it seems overkill
to use a polyfill.

The getSound() tooltip/picker is based on the image picker from webpages,
and that's where most of the code in this diff comes from.

Outside of this diff:
I will be adding an additional point to the program guidelines
about expected use of this functionality, and what we consider abuse.
Specifically, it will say that students should only use it
in response to user interaction, never inside a draw loop, and
never upon program load.
I will also send a subsequent diff for the varnish proxy to serve the sounds
off kasandbox.org.

Test Plan:
- Create program with playSound(), like:
```
var myObj = {
    x: 5,
    sound: getSound("rpg/giant-no")
};

var mousePressed = function() {
    myObj.x += 2;
    playSound(myObj.sound);
};
```
- Press the canvas and hear the sound.
- Type getSound(, see it autocomplete the default sound, see picker pop up, play new sound and press OK,
  now hear that new sound.
- Make sure that getImage( picker still works a before, type getImage(, see autocomplete, pick new image.
- Make sure that HTML image picker still works, by playing /computing/computer-programming/html-css/intro-to-html/p/html-images
- Run automated tests in live-editor (gulp test)

Reviewers: john

Reviewed By: john

Subscribers: brian, #cs

Differential Revision: https://phabricator.khanacademy.org/D15461