Move draggable error buddy logic into tipbar module.
I restricted the movement to only be in the vertical axis (there were only a few pixels of available horizontal movement, so it just made it janky). Additionally I limited the drag handle to be just the error buddy, to match the CSS cursor.

Test Plan:
Open `live-editor/demos/simple/` and generate an error in your program, have the error dialog come up. Verify that the error buddy is draggable and that you can move the dialog up-and-down (but not outside of the output window). Additionally the only drag handle should be the error buddy itself, not the entire window.

Auditors: pamela