Send links through a redirect URL if one is specified (live-editor diff)
Summary:
Previously, we attached listeners to links and showed users an alert
before letting them navigate to the links.
However, this was problematic in browsers with pop-up blockers (particularly
Safari, which has the pop-up blocker always on).
It's also problematic as it makes it harder to teach preventDefault for JS DOM course.
Instead, we do what Google Docs does, send URLs through a redirect interstitial
that requires them to click one more time.
We transform the URLs in live-editor in postProcessing of the webpage,
which is what this diff does.

Test Plan:
- Run tests (new tests in output_test.js)
- Put this HTML in /cs/new/webpage:
  http://pastebin.com/RXBvBNyD
- Click each of the links.
  The external link should go through the proxy redirect.
  The internal link should jump down the page.
  The KA link should go to it in a new window.
  The relative link should 404 in a new window.

Reviewers: john

Reviewed By: john

Subscribers: #cs

Differential Revision: https://phabricator.khanacademy.org/D15799