Disable brace autocompletion
Summary:
This is currently subjective, but the brace autocompletion (new line, enter, })
is fairly aggressive and takes a while to undo when it's accidental or unwanted.
We could technically measure the amount of undoing that happens after a brace
autocompletion but that's not super easy.
So my vote is for disabling it.

Test Plan:
- Type var x = {, notice it does not autocomplete
- Watch http://127.0.0.1:8080/computing/computer-programming/programming/objects/p/intro-to-objects,
  forward in 40 seconds, see that it still completes

Reviewers: john

Reviewed By: john

Subscribers: #cs

Differential Revision: https://phabricator.khanacademy.org/D16273