Add ability for some mp3s to be slower (for translations)
Summary:
Some languages have requested the ability to record the talk-throughs slowed down,
as we sometimes talk fast and other languages have longer words on average.
This does that in a lightweight way, by checking the filename for a multiplier
and then multiplying the command timestamps by that multiplier.
We can maybe do something similar to let people slow/speed up the English mp3s,
but tied to playbackRate instead.

Test Plan:
- Change getMp3URL in scratchpads.js to return "https://s3.amazonaws.com/ka-cs-scratchpad-audio/blob_x1.3.mp3";
  and watch http://127.0.0.1:8080/computing/computer-programming/html-css/intro-to-html/p/new-webpage,
  verify you can play/pause/seek, and see/hear expected output

Reviewers: john

Reviewed By: john

Subscribers: #cs

Differential Revision: https://phabricator.khanacademy.org/D16363