Show a message when runtime errors detected
Summary:
We were already showing a message when we found lint errors
in the JS in webpages. This change shows runtime messages
based on the onerror event on the window.
This might be too aggressive, like if there are runtime errors
not caused by the page. In that case, I can filter based on attributes
of the error like line number, source, type, etc.
The motivation for this change is the feedback from the HTML/JS challenges,
where many students are getting a ReferenceError but not realizing it.
I could also show the actual message up in the Oh Noes guy area,
but I'd like to encourage them to use the browser console more.

Test Plan:
- Create a new webpage, write:
    <script>
        bla(x);
    </script>
- See message "Your javascript encountered a runtime error. Check your console for more information."
- Comment it out, don't see it anymore

Reviewers: john

Reviewed By: john

Subscribers: #cs, kevinb

Differential Revision: https://phabricator.khanacademy.org/D16417