Use underscore to check inequality since objects may be compared
Summary:
I'm trying to make a challenge where the values need to be the same,
and its failing two results that are the same because its using the equality operator
on two objects. This makes it use _.equal instead.

Test Plan: - The 'Changing clothes' challenge step 2 should pass code as expected when true is passed in.

Reviewers: brian

Reviewed By: brian

Subscribers: #cs

Differential Revision: https://phabricator.khanacademy.org/D17350