Run jshint tests in pjs_output_test.
Summary:
Add back the jshint tests (from external/jshint/jshint_test.js)
to the pjs_output_test test. Two of the tests seem to time out when run
on the command line, but work when run in a browser. They don't time out
when only the jshint_test tests are run, however.

This is part of LintLove, so we can write linting tests for our changes.

Test Plan: - Run the tests (in the browser), see that they work.

Reviewers: john

Reviewed By: john

Subscribers: shannon

Differential Revision: https://phabricator.khanacademy.org/D18108