Call finishTest() for all tests.
Summary:
Firefox has a limit on the number of workers that can be created.
finishTest() kills any workers that have been created before running the next test,
but it wasn't being called for assertTest and allErrorsTest becuase I missed one
of the branches in the output.runCode callback.  This diff remedies this situation.

Test Plan: Run the tests in Firefox.  They should all pass and none should timeout.

Reviewers: emily

Subscribers: shannon

Differential Revision: https://phabricator.khanacademy.org/D18208