Prevent "string is not a function" when injecting code into OO programs
Summary:
When modifying the Asteroid blaster challenge a user will get an exception
that states "string is not a function".  This is because the instance
tracking code didn't have a reference to the array of instances.

https://github.com/Khan/live-editor/issues/279

This partial solves a similar bug: https://github.com/Khan/live-editor/issues/300.
You can make one modification successfully, but the next fails on the
next change you make.  I'll investigate this in a separate diff.

Test Plan: run tests/output/pjs/index.html

Reviewers: pamela

Reviewed By: pamela

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D19067