Avoid using gulp-handlebar's version of handlebars on fresh checkout.
Summary:
We use version 1.0-beta5 of handlebars in both live-editor and webapp.
When a developer does a fresh checkout of live-editor and installs the
npm dependencies, gulp-handlebars is installed which uses version 1.3
of handlebars.  By default version 1.3 is used to build the templates,
but we use the 1.0-beta5 runtime and the two don't interoperate.  To
solve this problem, the suggestion in the README.md is to delete the
gulp-handlebars version.  This diff automates that process.

Test Plan:
- rm -rf node_modules/gulp-handlebars
- npm install
- rm -rf build/tmpl
- gulp build
- git add build/tmpl
- git status should show no files have changed

Reviewers: john, pamela

Reviewed By: pamela

Subscribers: pamela

Differential Revision: https://phabricator.khanacademy.org/D19077