Cleanup lint errors picked up when updating substate in webapp
Summary:
Since adding a lint_blacklist.txt to live-editor, khan-linter now uses
it when updating substate.  In order to make the linter happy we blacklist
build/, demos/, external/ and tests/.  We also ignore a couple of files:
- processing-stubs.js (has hundreds of lines of line)
- sql-results.handlebars (requires i18n-izing)

I will circle back and fix the two individual files in a subsequent diff.

Eventually, I'd like to fix the lint in tests/ as well.

Test Plan:
- commit third_party/javascript-khansrc/live-editor
- ../devtools/khan-linter/runlint.py --blacklist=yes third_party/javascript-khansrc/live-editor

Reviewers: jlfwong

Reviewed By: jlfwong

Subscribers: john, pamela

Differential Revision: https://phabricator.khanacademy.org/D19109