Check that we have all of the source files and deps listed in build-paths.json before building.
Summary:
This is the first step in fixing the issue where code in the build/ folder
is out of sync with the source code in the js/ folder.  The problem is, is
that someone can make a change to a file and forget to build it.  If that
change is checked in, as long as no one touches that file again, that change
will not make it into production because the build system only builds things
when the source changes.

Another situation which one of our contributors got into is he was missing
some of dependencies.  When he commit the built files there was lots of code
missing.  Thankfully we noticed the issue, but if there are changes to the
deps it can be hard to see what's changing in the built files.

This diff checks to make sure that we have all of the deps before building.
A future diff will clean the build folder before building.  The performance
of subsequent builds should not be affected by this when using the watcher.

https://github.com/Khan/live-editor/issues/322

Test Plan:
- delete folers from bower_components and external
- run "gulp"
- verify that it stops the build and prints out which files are missing

Reviewers: john, pamela

Reviewed By: pamela

Subscribers: pamela

Differential Revision: https://phabricator.khanacademy.org/D19078