Run iframe tests using phantomjs
Summary:
tests/output/pjs/index.html runs its suite of tests inside an iframe so
that we can exercise code in pjs-output.js that relies on being inside
an iframe.  gulp-mocha-phantomjs uses sockets to communicate with mocha
in the browser and doesn't work across excution contexts such as iframes.

The solution is to use the same json reporter strategy that webapp.  This
has been modified a bit so that we can access results from node.  It
consists of:
- a custom reporter which prints test result messages to the console in the browser
- a script which runs in phantomjs which echoes the messages from the browser to stdout
- a gulp plugin which executes the phantomjs script and reads the messages from stdout
  and parses them to produce test results

{F155035}

Test Plan: - `gulp test` should run all tests including the tests/output/pjs/index.html

Reviewers: pamela, john

Reviewed By: john

Subscribers: john, jlfwong

Differential Revision: https://phabricator.khanacademy.org/D19123