Force rebuild files before watching.
Summary:
Summary
The "scripts" and "styles" tasks were only building files
if the source files are newer than the corresponding destination
files.  If someone forgets to commit the build files and
someone else does a fresh checkout even though some source
files have changed they won't be included in the build.  This
forces all files to be built the first time "script" and "styles"
are run.  Subsequent builds use gulp-newer like before so
that we're only compiling files that need to be.

Test Plan:
- run `gulp build`
- verify that files in build/ have changed

Reviewers: john, pamela

Reviewed By: pamela

Subscribers: pamela

Differential Revision: https://phabricator.khanacademy.org/D19128