Disable	auto-close/auto-fill during recording playback/seeking.
Summary:
https://github.com/Khan/live-editor/issues/349
All text that was typed	and auto-filled	when recording a talk-through
gets saved as commands.	When playing back those commands if auto-fill
is enabled the auto-filled text	can remain when pausing or seeking b/c
the text is essentially	added twice: once from the commands and	once
from autofill. In order to correct this we disable auto-fill during
playback and seek operations.

Test Plan:
- run "gulp test_record" to grab recording-data.json from S3
- run "tests/record/index.html", re-run any failed suites in isolation
- create a new pjs recording and type the following text
```
   rect (100, 100, 100, 100);
   draw = function() {
       fill(mouseX, mouseY, mouseY);
   };
```
- play the recording back and pause after the start of the `fill` command
- jump to various locations and make sure that there's never any extra text like:
```
   fill(mouseX, mouseY, mouseY);255, 0, 0);
```

Reviewers: pamela

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D19211