Make createGraphics as a "safe" call and rerun if there are any calls to createGraphics.
Summary:
The reason why we have to rerun whenever createGraphics is present
is not actually createGraphics, but rather calls that render stuff
to the Processing instances returned by createGraphics.  In the
future we might be able to reuse these instances, but we'd need
to track which call to createGraphics returned which instance.  See
the code for additional comments.

Test Plan:
- create a pjs program with the following code
```
    var img = getImage("avatars/leafers-seedling");

    var gfx = createGraphics(400, 400, 1);
    gfx.image(img, 37, 0);

    draw = function() {
        background(0, 0, 0);
        image(gfx, 10, 0);
    };
```
- verify that the gray rect with the seedling can be moved via the scubber
- verify that the seedling can be moved within the gray rect via the scrubber

Reviewers: john, pamela

Reviewed By: john, pamela

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D19256