Prevent babel from killing the watcher
Summary:
Gulp's watcher stops if any of the plugins throw an exception which
happens if a developer saves a source file with a syntax error.  I
tried using .on('error', handler) and gulp-plumber.  These solutions
kept the watcher alive, but for some reason the babel plugin stops
working after the first error, even if the syntax error is fixed.

I created a copy of the gulp-babel plugin and modified it so that
it logs errors instead of throwing them.  This fixes the issue, but
has the undesired side effect that the build file will be without
the file containing the syntax error.  When the tests run almost all
of them will fail because of this.  The output is kind of spammy so
I'm planning to truncate the output of the test runner in a following
diff.

This is still not ideal because it shouldn't be running the tests
at all.  Also, the tests take a while to run.  If someone makes a
change and everything compiles then we should stop the tests that
are currently being run and start running the tests on the new code.
I have started working on something that will eventually address
these issues, but since they aren't super high priority I'm just
working on it in my spare time.

Test Plan:
- run 'gulp'
- make a syntax error
- verify that gulp logs the error (and runs the tests w/ lots of failures)
- fix the error
- verify that gulp compiles the code and runs the tests with lots of passes

Reviewers: john

Reviewed By: john

Subscribers: pamela

Differential Revision: https://phabricator.khanacademy.org/D19377