Check that a sound is in our list of sounds before trying to load it.
Summary:
phantomjs was timing out when trying to load sounds that don't exist.  It
also doesn't make sense to try and load sounds that we know don't exist.
Instead, we check if a sound is in OutputSounds before trying to load it.
We should probably do the same for images, but I've left that for a future
diff.

Test Plan: - gulp test_output_pjs, verify that there are no failures

Reviewers: pamela

Reviewed By: pamela

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D19365