Adds in a new document mode and textarea editor.
Summary:
This adds in a new "document" mode to the editor. It's very simple, it has no output whatsoever (so there actually isn't a 'document' output module). Additionally I've added in a new "textarea" editor which is like a very simple version of Ace (but it supports non-monospaced fonts!). We're going to be using this for the new writing-based projects on KA.

Most of the changes are simple, the biggest are:

- We no longer require "execFile" and when it's not passed in we actually hide the output area (these are the changes to Handlebars and the CSS).
- We have a new textarea editor mode. It handles most of the basic stuff properly (get/set text, get/set cursor).
- Basic text screenshot support has been added to live-editor.js. This just takes the text from the user's program and turns it into a basic screenshot (See below).

Example of screenshots in use:
{F157558}

Test Plan: The test suite runs the same as before! Honestly not sure what we'd test - this new mode is sort of the absence of any functionality. Suggestions recommended!

Reviewers: kevinb, pamela, meredith

Subscribers: #cs

Differential Revision: https://phabricator.khanacademy.org/D19396