Fix when tooltips appear.
Summary:
I errorneously change the tooltip precedence in	a previous commit when
trying to correct a problem where the number scrubber wasn't showing up.
The real reason	it wasn't showing up is because ace editor now send
'cursorChanged'	events after the user clicks to the change the cursor location.
This is perfectly reasonable, the only problem is we get two events when
the user clicks, a click and cursorChanged event.  In order to filter out
that second event we track when the mouse button is down.  It just so
happens that cursorChanged events happen between mousedown and mouseup so
this works.

Test Plan: - run tests/tooltips/index.html

Reviewers: pamela

Reviewed By: pamela

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D19414