Give pjs more time to execute before the LoopProtector throws
Summary:
Some existing sketchpads do lots of processing and may trigger
the pjs LoopProtector.  This diff makes the timeout for the pjs
environment a little longer.  Since the LoopProtector is time
based, some programs that work on certain machines may stop working
on slower machines.  This diff does not address this problem.
We might want to try doing some sort of benchmarking with the
program initially starts.

Test Plan:
- load /computer-programming/metaballs/6209526669246464 and try adjust
the PX_SIZE.  Oh noes guy should stop the program when PX_SIZE is 2 +/- 1

Reviewers: pamela

Reviewed By: pamela

Differential Revision: https://phabricator.khanacademy.org/D19462