Prevent KAInfiniteLoop functions from being injected and don't delete them either.
Summary:
https://app.asana.com/0/28126167511849/41848983568884
When deleting all of the code, inject would add delete statements which would cause
KAInfiniteLoopSetTimeout to be undefined at the end.  After fixing that, and undoing,
there would be extra code inject to add the KAInfiniteLoop functions back because
they were being grabbed as global functions so we ignore those in a couple of places.
The reason why these are showing up is being these functions are being added to the
global context/canvas object.

Test Plan: - load tests/output/pjs/index.html and run the tests

Reviewers: pamela

Reviewed By: pamela

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D19623