Report location of long running loops.
Summary:
The LoopProtector in the pjs environment has been upgrade to
port the location of long running loops.  An option has been added
to the LoopProtector constructor called 'reportLocation'.  If this
option is true then each call to KAInfiniteLoopProtect() is passed
a JSON string containing information about the location of the this
call.  Each time KAInfiniteLoopProtect is called with this string
it records a count.  If the timeout is trigger it counts up how many
times each the call was made at each callsite and reports the one
that's been called the most.

Test Plan:
- run the automated tests
- copy/paste the following code into the live-editor
```
    fill(255,0,0);
    mouseClicked = function() {
        ellipse(mouseX, mouseY, 100, 100);
        while (true) {

        }
    };
```
- verify that the Oh noes! guy pops up and says it's a "while" loop
- verify that click on "show me where" takes you the line with the "while" loop

Reviewers: pamela, john

Reviewed By: pamela, john

Subscribers: john

Differential Revision: https://phabricator.khanacademy.org/D19639