Stop running pjs code in a worker before running it in the iframe.
Summary:
The whole reason we were running pjs user code in a worker was to
detect infinite loops.  Now that the pjs environment is using the
LoopDetector we no longer need to run user code in a web worker.
The result of this is that we can get rid of worker.js as well as
the stub files.  The infinite loop tests needed to be changed.  We
now wait for the loopProtector's callback to fire.

Test Plan: - load tests/output/pjs/index.html

Reviewers: john, pamela

Reviewed By: john, pamela

Differential Revision: https://phabricator.khanacademy.org/D19781