Use LoopProtector's callback for all infinite loop tests.
Summary:
One of the tests was catching an async exception which is super awkward to
do so we'll use LoopProtector's callback just like the other infinite
loop tests.  This means we can ditch the custom build of mocha.js.  I've
add some comments describing why we need to throw an exception even if
we don't end up catching it, though we probably should have a global
exception handler so we're not cluttering up people's consoles, but this
is really a nice to have so I'll leave it for now.

Test Plan: - gulp test_output_pjs

Reviewers: pamela, john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D19824