Make sure that infinite loops in main program body are reported with location information
Summary:
The loop errors being reported through runCode's callback and LoopProtector's were
slightly different.  Now we create the error object with the correct text in LoopProtector
and we throw/pass to the callback the same object now so we get the same message
regardless of which code path it takes.

Test Plan:
- gulp test_output_pjs
- gulp test_webpage_pjs (make sure LoopProtector still functions there as well)

Reviewers: pamela, john

Reviewed By: john

Differential Revision: https://phabricator.khanacademy.org/D19874